Relicense zipio files under the MIT license
Zipio files were licensed under a restrictive license containing a clause
mandating that any modifications to the source code must be sent via e-mail
to the copyright owner. The author was contacted and permitted the files
to be relicensed under the MIT license. This should avoid any problems
restributing FIGlet as free software in Linux distributions.

Signed-off-by: Claudio Matsuoka <cmatsuoka@gmail.com>